+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   * Included in static site
 * SQLite database has foreign keys
 * New CLI command "versionedbuild" - currently used for internal testing
+* New field type, "integer"
 
 ### Changed